alterações cad-usuario.jsp

+++ src/main/java/dao/DAOUsuarioRepository.java
@@ -77,10 +77,10 @@ public class DAOUsuarioRepository {
 	}
 	
 	
-	public ModelLogin consultaUserId(String id) throws Exception {
+	public ModelLogin consultaUsuarioID(String id) throws Exception {
 
 		ModelLogin modelLogin = new ModelLogin();
-		String sql = "select * from model_login where id= ?";
+		String sql = "select * from model_login where id= ? and useradmin is false";
 		PreparedStatement stat = connection.prepareStatement(sql);
 		stat.setLong(1, Long.parseLong(id));
 		
@@ -133,7 +133,7 @@ public class DAOUsuarioRepository {
 	
 	public void deletarUser (String idUser) throws Exception {
 		
-		String sql = "delete from model_login where id = ? ;";
+		String sql = "delete from model_login where id = ?  and useradmin is false ;";
 		PreparedStatement prepararSql = connection.prepareStatement(sql);
 		prepararSql.setLong(1, Long.parseLong(idUser));
 		prepararSql.executeUpdate();
@@ -142,11 +142,37 @@ public class DAOUsuarioRepository {
 		
 	}
 	
+public List<ModelLogin> consultaUsuarioList() throws Exception {
+		
+		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
+		
+		String sql = "select * from model_login where useradmin is false";
+		PreparedStatement statement = connection.prepareStatement(sql);		
+		
+		ResultSet resultado = statement.executeQuery();
+		
+		while (resultado.next()) { /*percorrer as linhas de resultado do SQL*/
+			
+			ModelLogin modelLogin = new ModelLogin();
+			
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setNome(resultado.getString("nome"));
+			//modelLogin.setSenha(resultado.getString("senha"));
+			
+			retorno.add(modelLogin);
+		}
+		
+		
+		return retorno;
+	}
+	
 public List<ModelLogin> consultaUsuarioList(String nome) throws Exception {
 		
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 		
-		String sql = "select * from model_login  where upper(nome) like upper(?) ";
+		String sql = "select * from model_login  where upper(nome) like upper(?) and useradmin is false";
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setString(1, "%" + nome + "%");
